Add files via upload
This file contains all of the items on the food traceability list (FTL). There is a header, so ensure you account for that when loading it into a python file. 

The list contains about an equal number of foods per category. Therefore, there are many duplicates. This is in order to ensure that each category has the same chance of being selected as a food.